% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestSubsetSel.R
\name{bestSubsetSel}
\alias{bestSubsetSel}
\title{bestSubsetSel - best subset selection}
\usage{
bestSubsetSel(
  dat,
  respVar,
  method,
  modelString = NULL,
  family = NULL,
  saveOutput = NULL
)
}
\arguments{
\item{dat}{a data frame containing the covariates under consideration as well as the response variable. Do not include superfluous variables.}

\item{respVar}{the name of the response variable in dat (just the var name, no need to use a string)}

\item{method}{the type of model to fit, either "lm" or "glm"}

\item{modelString}{an additional string to include in all of the model formula. This is useful for forcing certain variables into the model. Must be set up with a leading plus sign like "+varName". If theres nothing you want to force in, this should be NULL. Defaults to NULL. This does end up having the same term in the model multiple times in some of the fromula but this is not an issue, the term is still only estimated once. You can check with summary(lm(Sepal.Length ~ Sepal.Width+Sepal.Width, data = iris))}

\item{family}{distribution family for glm model. If not using method glm, then this should be NULL. Defaults to NULL}

\item{saveOutput}{a file path to a directory in which to save the outputs. Especially helpful if this is going to take a long time and you want to be able to pull the results in without having to rerun this function. It will save it in the specified directory named with the date and time. If you do not want to save the results, then this should be NULL. Defaults to NULL.}
}
\value{
a list of the formula for all of the models fit, their associated IC, as well as some of the input arguements. Returns an object of class "bss" so this can be used with my summary.bss() s3.
}
\description{
bestSubsetSel - best subset selection
}
\details{
this function fits a model for all possible variable combinations and returns the AIC and BIC. In the future, it would be nice to set this up for more types of models beyond lm and glm like mixed effects models as well as adding in different ICs.
}
\examples{
data("iris")

lmBss <- bestSubsetSel(dat = iris, respVar = Sepal.Length,
 method = "lm", modelString = "+Sepal.Width")
summary(lmBss)

irisBin <- iris |> 
dplyr::mutate(Species = ifelse(Species == "virginica", 0, 1))


glmBss <- bestSubsetSel(dat = irisBin, respVar = Species,
 method = "glm", family = "binomial", modelString = "+Sepal.Width")
summary(glmBss)
}
