% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomScoreTest.R
\name{binomScoreTest}
\alias{binomScoreTest}
\title{binomScoreTest - binomial score test}
\usage{
binomScoreTest(y, n, pi_hy = 0.5, alpha = 0.05)
}
\arguments{
\item{y}{test}
}
\value{
A list containing values:
\code{ML_estimate} test1,
\code{test_stat} test2,
\code{p_val},
\code{ci}
}
\description{
binomScoreTest - binomial score test
}
\details{
some explaination
}
\examples{
#test
}
