% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyColors.R
\name{easyColors}
\alias{easyColors}
\title{easyColors - Pallet of N Colors}
\usage{
easyColors(n)
}
\arguments{
\item{n}{number of colors to return}
}
\value{
A list containing values:
\code{resultHolder} placeholder
}
\description{
easyColors - Pallet of N Colors
}
\details{
I believe this is how ggplot chooses colors for discrete variables, it is \code{n} equally spaced colors based on rgb value
}
\examples{
#test
}
