% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqFormat.R
\name{freqFormat}
\alias{freqFormat}
\title{freqFormat - Formatting Univariate Frequency Tables}
\usage{
freqFormat(x, varName = NULL, extraText = NULL)
}
\arguments{
\item{x}{an output of a frequency table from uniFreq() or unifreq2()}

\item{varName}{the desired name of the variable as a string}

\item{extraText}{extra text to add into the title}
}
\value{
a formatted univariate frequency table
}
\description{
freqFormat - Formatting Univariate Frequency Tables
}
\details{
function for formatting the frequency tables caclulated with uniFreq and uniFreq2 functions
with uniFreq: pass into x the item from the list to be formatted, use double bracket subsetting. this is done for one single element of uniFreq output at a time
can also call names on the uniFreq object to know which number to use in bracket
for unifreq: to display all of the tables you want, use a for loop in r markdown but make sure that you have results='asis' in the header otherwise it prints a bunch of crap
}
\examples{
#test
}
