% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupFreq.R
\name{groupFreq}
\alias{groupFreq}
\title{groupFreq - Frequency Table for Multiple Variables}
\usage{
groupFreq(data, variable, groupings, includeNA = TRUE)
}
\arguments{
\item{data}{the data set}

\item{variable}{a string with the name of the main variable of interest}

\item{groupings}{a vector of strings with the name of the variables interested in grouping on}

\item{includeNA}{whether or not to include missing values in the frequency table, default is \code{TRUE}}
}
\value{
a grouped frequency table
}
\description{
groupFreq - Frequency Table for Multiple Variables
}
\details{
create a frequency table based on a grouping
like uniFreq2(), this creates one table at a time
this function has a different formatting function than uniFreq() and uniFreq2()
THIS IS SET UP FOR ONLY TWO VARIABLES
ie ONE VARIABLE FOR variable ARGUMENT, ONE VATIABLE FOR groupings ARGUEMENT
IF YOU WANT TO UPDATE THIS FUNCTION TO TAKE MORE GROUPINGS YOU NEED TO ADD MORE UNGROUPING AND GROUPING ARGUEMENT
TO GET THE cumulative frequency and cumulative percent correct
}
\examples{
#test
}
