% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupFreqForm.R
\name{groupFreqForm}
\alias{groupFreqForm}
\title{groupFreqForm - Formatting Grouped Frequency Tables}
\usage{
groupFreqForm(x, varNames = NULL, extraText = NULL)
}
\arguments{
\item{x}{an output of a frequency table from groupFreq()}

\item{varNames}{a vector of strings that are the desired presentation names of the variables}

\item{extraText}{any additional things you want to add to title}
}
\value{
a formatted grouped frequency table
}
\description{
groupFreqForm - Formatting Grouped Frequency Tables
}
\details{
formatting function for groupFreq()
THIS SHOULD PROBABLY BE UPDATED
}
\examples{
#test
}
