% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kablePdf.R
\name{kablePdf}
\alias{kablePdf}
\title{kablePdf - Formatting a Table for Markdown PDF}
\usage{
kablePdf(table, title = "title", colNames = NULL, digits = NULL)
}
\arguments{
\item{table}{the table to format, can be piped}

\item{title}{a title for the table, default is "title"}

\item{colNames}{a vector of strings containing new names for the columns}

\item{digits}{the number of digits to display}
}
\value{
a formatted table
\item{retTab}{a formatted table}
}
\description{
kablePdf - Formatting a Table for Markdown PDF
}
\details{
Basic table formatting for a markdown pdf document
}
\examples{
#test
}
