% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagCountSurvival.R
\name{lagCountSurvival}
\alias{lagCountSurvival}
\title{lagCountSurvival - creating the start stop survival for the lag time varying covariate}
\usage{
lagCountSurvival(
  x,
  visitVar = recallTimes,
  failVar = fail,
  idVar = id,
  toothVar = tooth,
  windowWidth = 365,
  splitChar = "\\\\|"
)
}
\arguments{
\item{x}{data formatted like in example}

\item{visitVar}{the name of the time varying variable in x to count (just the var name, no need to use a string)}

\item{failVar}{the name of the variable in x holding the failure information (just the var name, no need to use a string)}

\item{idVar}{the name of the variable in x that has ID information (just the var name, no need to use a string)}

\item{toothVar}{the name of the variable in x that has the tooth number information, in this use case, observations are uniquely identified thru id and tooth number but in other applications only one ID would be needed and this would have to be augmented (just the var name, no need to use a string)}

\item{windowWidth}{lag period to count in}

\item{splitChar}{the regular expression for the character that splits the visitVar}
}
\value{
formatted start stop survival data with lag covariate
}
\description{
lagCountSurvival - creating the start stop survival for the lag time varying covariate
}
\details{
function for creating the intervals for the time varying covariate which is counts in a lag window, in this case number of recall appointments in 365 days this is set up for a particular analysis (isolation methods satoki) so the format may not be directly transferable to other data sets but with slight tweeks, it  can be made to fit, see the example data above. "recallTimes" is the amount of days after initial restoration that the patient had an appointment, stored as a character vector with specific format
}
\examples{
datToy <- data.frame(id = c(1,1,2),
      tooth = c(4,5,4),
      fail = c(0,1,1),
      recallTimes = c("0|400",
                "0|180|365|450|550|720|1460|1560",
                "0|180|360|1440|1540"),
      sex = c("M", "M", "F"))
lagCountSurvival(x=datToy)
}
