% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagWindow.R
\name{lagWindow}
\alias{lagWindow}
\title{lagWindow - counting number of elements in moving window}
\usage{
lagWindow(x, windowWidth = 365)
}
\arguments{
\item{x}{the variable to work with}

\item{windowWidth}{size of the window}
}
\value{
a vector of counts in a window
}
\description{
lagWindow - counting number of elements in moving window
}
\details{
this is designed to just be a helper function for lagCountSurvival(). I know that there is a way to make it so this function is not available to the user and only available to code within this package but i dont want to do that right now. i believe all you need to do is remove the export command but im cool with this for now
}
\examples{
#see lagCountSurvival()
}
