% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percForm.R
\name{percForm}
\alias{percForm}
\title{percForm - Formatting Percentages as Strings}
\usage{
percForm(x)
}
\arguments{
\item{x}{values or values taking value 0-100 to be formatted}
}
\value{
a string formatted as a percentage with the percentage symbol
}
\description{
percForm - Formatting Percentages as Strings
}
\details{
Formatting percentages as strings with 2 decimal places and the percentage symbol. This is nice when making tables and some of my univariate table functions use this function.
}
\examples{
#test
}
