% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postDenMode.R
\name{postDenMode}
\alias{postDenMode}
\title{postDenMode - calculate posterior density mode from posterior samples}
\usage{
postDenMode(pd)
}
\arguments{
\item{pd}{a vector of posterior draws}
}
\value{
the posterior density mode
}
\description{
postDenMode - calculate posterior density mode from posterior samples
}
\details{
Calculate posterior density mode from posterior samples, this is often perfered over the posterior mean but in case of symmetry it will be about the same. For continuous variables, this translates to the value associated with the max density in the posterior. This has an if clause in it for when all the posterior draws are the same ie when you are doing some sort of imputation task and the value does not need to be estimated bc it is known.
}
\examples{
#placeholder
}
