% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probDirect.R
\name{probDirect}
\alias{probDirect}
\title{probDirect - probability of direction for posterior samples}
\usage{
probDirect(pd)
}
\arguments{
\item{pd}{a vector of posterior draws}
}
\value{
the probability of direction
}
\description{
probDirect - probability of direction for posterior samples
}
\details{
What is the probability that the parameter is positive/negative? The sign is based on the posterior density mode, ie if the posterior density mode is positive, this is the probability that the parameter is positive.
}
\examples{
#placeholder
}
