% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop2SampN.R
\name{prop2SampN}
\alias{prop2SampN}
\title{prop2SampN - Sample size calculation for 2 sample proportion test}
\usage{
prop2SampN(p1, p2, alpha_, power)
}
\arguments{
\item{p1}{assumed proportion in group 1}

\item{p2}{assumed proportion in group 2}

\item{alpha_}{desired alpha level}

\item{power}{desired power}
}
\value{
a list containing
\code{conditionalN} the sample size needed to detect specified difference at desired power level with desired alpha for the conditional test
\code{unconditionalN} the sample size needed to detect specified difference at desired power level with desired alpha for the unconditional test
}
\description{
prop2SampN - Sample size calculation for 2 sample proportion test
}
\details{
This test is for two sided tests only, with n being the sample size for each group and assuming equal allocation. For details on the calculations, see clinical trials notes.
}
\examples{
prop2SampN(p1 = 0.16, p2 = 0.10, alpha_ = 0.05, power = .85)
}
