% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssSummaryS3.R
\name{summary.bss}
\alias{summary.bss}
\title{summary.bss - summary function for "bss" objects}
\usage{
\method{summary}{bss}(object, toReturn = NULL, n = NULL)
}
\arguments{
\item{object}{a bss object created by bestSubsetSel()}

\item{toReturn}{the name of the IC to return. Default is NULL and it returns all of the ICs. Options are "aic" and "bic" right now}

\item{n}{number of top models to return for each IC that is being returned}
}
\value{
a table of the top models in order of lowest IC to highest IC. Each model formula is given as a string, not a formula, so you cant run something like "lm" on that columns contents. However, each model is given with its associated index from the bestSubsetSel() output so you can trace it back in that ouput.
}
\description{
summary.bss - summary function for "bss" objects
}
\details{
summary for "bss" objects
}
\examples{
data("iris")

lmBss <- bestSubsetSel(dat = iris, respVar = Sepal.Length,
 method = "lm", modelString = "+Sepal.Width")
summary(lmBss)

irisBin <- iris |> 
dplyr::mutate(Species = ifelse(Species == "virginica", 0, 1))


glmBss <- bestSubsetSel(dat = irisBin, respVar = Species,
 method = "glm", family = "binomial", modelString = "+Sepal.Width")
summary(glmBss)
}
