% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniFreq.R
\name{uniFreq}
\alias{uniFreq}
\title{uniFreq - Frequency Table for all Variables}
\usage{
uniFreq(x)
}
\arguments{
\item{x}{a data.frame where each row represents an observation}
}
\value{
a list of univariate frequency tables, one for each variable in the data set
}
\description{
uniFreq - Frequency Table for all Variables
}
\details{
This function creates a univariate frequency table for each variable in the data set. I generally tend to use uniFreq2 more since this does all variables and that sometimes is a little much.
}
\examples{
#test
}
