% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniFreq2.R
\name{uniFreq2}
\alias{uniFreq2}
\title{uniFreq2 - Frequency Table for a Single Variable}
\usage{
uniFreq2(data, variable, includeNA = TRUE)
}
\arguments{
\item{data}{the name of the data frame}

\item{variable}{the name of the variable stored as a character}

\item{includeNA}{whether or not to include missing values in the frequency table, default is \code{TRUE}}
}
\value{
a univariate frequency table
}
\description{
uniFreq2 - Frequency Table for a Single Variable
}
\details{
create a frequency table for a single variable at a time.
can be formatted with the freqFormat function.
a potential issue here is if you use this with a numeric variable it will probably have all the factor levels in a random order.
to combat this, could add an if statement to this code to specifically handle turning numeric to factor and getting the order of the levels right.
}
\examples{
#test
}
