% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniFreqForm.R
\name{uniFreqForm}
\alias{uniFreqForm}
\title{uniFreqForm - Formatting Univariate Frequency Tables}
\usage{
uniFreqForm(x, varName = NULL, extraText = NULL)
}
\arguments{
\item{x}{an output of a frequency table from uniFreq() or unifreq2()}

\item{varName}{the desired name of the variable as a string, default is NULL
and defaults to name of the variable in the table. The supplied name or the
name in the table cannot contain an underscore (probably other special
characters too).}

\item{extraText}{extra text to add into the title}
}
\value{
a formatted univariate frequency table
}
\description{
uniFreqForm - Formatting Univariate Frequency Tables
}
\details{
Function for formatting the frequency tables caclulated with uniFreq
and uniFreq2 functions. Formats to latex for outputting in R markdown pdf docs \n
This function operates on a single table at a time. That means if you have
used uniFreq and have a list of output tables, you need to loop/lapply
through them and apply this formatting function to one table at a time. \n
If you have a list of uniFreqForm outputs and want to have them all output
in an R markdown, make sure the chunk has results="asis" in the chunk header.
}
\examples{
#test
}
