% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomScoreTest.R
\name{binomScoreTest}
\alias{binomScoreTest}
\title{binomScoreTest - binomial score test}
\usage{
binomScoreTest(y, n, pi_hy = 0.5, alpha = 0.05)
}
\arguments{
\item{y}{test}
}
\value{
A list containing values:
\item{ML_estimate}{test1}
\item{test_stat}{test2}
\item{p_val}
\item{ci}
}
\description{
binomScoreTest - binomial score test
}
\details{
some explaination
}
\examples{
#test
}
