% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkGammaTest.R
\name{gkGammaTest}
\alias{gkGammaTest}
\title{gkGammaTest - Goodman Kruskal Gamma Test}
\usage{
gkGammaTest(x)
}
\arguments{
\item{x}{test}
}
\value{
A list containing values:
\code{resultHolder} placeholder
}
\description{
gkGammaTest - Goodman Kruskal Gamma Test
}
\details{
need to do the documentation for this one, see the actual code for commented instructions
}
\examples{
#test
}
