% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kablePdf.R
\name{kablePdf}
\alias{kablePdf}
\title{kablePdf - Formatting a Table for Markdown PDF}
\usage{
kablePdf(table, title = "title", colNames = NULL, digits = NULL)
}
\arguments{
\item{table}{the table to format, can be piped}

\item{title}{a title for the table, default is "title"}

\item{colNames}{a vector of strings containing new names for the columns}

\item{digits}{the number of digits to display}
}
\value{
a formatted table
\item{retTab}{a formatted table}
}
\description{
kablePdf - Formatting a Table for Markdown PDF
}
\details{
Basic table formatting for a markdown pdf document. This gets things set up
in a standard way that I like, but can be further adjusted by piping in to
further kableExtra functions. I havent full explored all of the options, but
you can at least do some things. See examples
}
\examples{
#piping custom table into further kableExtra functions for further customization
head(iris) |>
  kablePdf(title = "Iris Example") |>
  #change column 3 justificiation to left
  column_spec(3, latex_column_spec = "l")
}
